--- a/holidays/countries/botswana.py
+++ b/holidays/countries/botswana.py
@@ -16,8 +16,7 @@ from dateutil.easter import easter
 from dateutil.relativedelta import MO
 from dateutil.relativedelta import relativedelta as rd
 
-from holidays.constants import JAN, MAY, JUL, SEP, OCT, DEC, SAT, SUN
-from holidays.constants import HOLIDAY_NAME_DELIMITER
+from holidays.constants import JAN, MAY, JUL, SEP, DEC, SAT, SUN
 from holidays.holiday_base import HolidayBase
 
 
@@ -33,67 +32,62 @@ class Botswana(HolidayBase):
     special_holidays = {2019: ((JUL, 2, "Public Holiday"),)}
 
     def _populate(self, year: int):
+        def _add_with_observed(
+            hol_date: date, hol_name1: str, hol_name2: str = ""
+        ) -> None:
+            self[hol_date] = hol_name1
+            if hol_name2:
+                self[hol_date + td(days=+1)] = hol_name2
+            if self.observed and year >= 1995:
+                if hol_date.weekday() == SUN:
+                    obs_date = hol_date + td(days=2 if hol_name2 else 1)
+                    self[obs_date] = f"{hol_name1} (Observed)"
+                elif hol_date.weekday() == SAT and hol_name2:
+                    obs_date = hol_date + td(days=+2)
+                    self[obs_date] = f"{hol_name2} (Observed)"
+
         if year <= 1965:
             return
         super()._populate(year)
 
-        self[date(year, JAN, 1)] = "New Year's Day"
-        self[date(year, JAN, 2)] = "New Year's Day Holiday"
+        _add_with_observed(
+            date(year, JAN, 1), "New Year's Day", "New Year's Day Holiday"
+        )
 
         # Easter and easter related calculations
         easter_date = easter(year)
         self[easter_date + td(days=-2)] = "Good Friday"
         self[easter_date + td(days=-1)] = "Holy Saturday"
         self[easter_date + td(days=+1)] = "Easter Monday"
-
-        self[date(year, MAY, 1)] = "Labour Day"
         self[easter_date + td(days=+39)] = "Ascension Day"
 
-        self[date(year, JUL, 1)] = "Sir Seretse Khama Day"
+        _add_with_observed(date(year, MAY, 1), "Labour Day")
+        if (
+            self.observed
+            and year >= 2016
+            and date(year, MAY, 1).weekday() == SAT
+        ):
+            self[date(year, MAY, 1) + td(days=+2)] = "Labour Day Holiday"
+
+        _add_with_observed(date(year, JUL, 1), "Sir Seretse Khama Day")
 
         # 3rd Monday of July = "President's Day"
-        d = date(year, JUL, 1) + rd(weekday=MO(+3))
-        self[d] = "President's Day"
-        self[d + td(days=+1)] = "President's Day Holiday"
+        dt = date(year, JUL, 1) + rd(weekday=MO(+3))
+        self[dt] = "President's Day"
+        self[dt + td(days=+1)] = "President's Day Holiday"
 
-        self[date(year, SEP, 30)] = "Botswana Day"
-        self[date(year, OCT, 1)] = "Botswana Day Holiday"
+        _add_with_observed(
+            date(year, SEP, 30), "Botswana Day", "Botswana Day Holiday"
+        )
 
-        self[date(year, DEC, 25)] = "Christmas Day"
-        self[date(year, DEC, 26)] = "Boxing Day"
+        _add_with_observed(date(year, DEC, 25), "Christmas Day", "Boxing Day")
 
-        if self.observed:
-            for k, v in list(self.items()):
-                # Whenever Boxing Day falls on a Saturday,
-                # it rolls over to the following Monday
-                if (
-                    2016 <= year == k.year
-                    and k.weekday() == SAT
-                    and v.upper() in {"BOXING DAY", "LABOUR DAY"}
-                ):
-                    # Add the (Observed) holiday
-                    self[k + td(days=+2)] = v + " Holiday"
-                if (
-                    1995 <= year == k.year
-                    and k.weekday() == SUN
-                    and v.upper() != "NEW YEAR'S DAY HOLIDAY"
-                ):
-                    # Add the (Observed) holiday
-                    self[k + td(days=+1)] = v + " (Observed)"
-
-                # If there is a holiday and an (Observed) holiday
-                # on the same day, add an (Observed) holiday for that holiday
-                hol_names = self.get(k).split(HOLIDAY_NAME_DELIMITER)
-                if len(hol_names) > 1:
-                    # self.get(date) returns a string containing holidays as a
-                    # comma delimited string split on delimiter to determine if
-                    # there are multiple on the same day
-                    # Add an (Observed) for the one that is not (Observed)
-                    for name in hol_names:
-                        if " (Observed)" not in name:
-                            self[k + td(days=+1)] = (
-                                name.lstrip() + " (Observed)"
-                            )
+        if (
+            self.observed
+            and year >= 2016
+            and date(year, DEC, 26).weekday() == SAT
+        ):
+            self[date(year, DEC, 26) + td(days=+2)] = "Boxing Day Holiday"
 
 
 class BW(Botswana):