Observed holidays calculation method changed. Here are some performance tests.
```
def calc_hol1():
    for year in range(1966, 2100):
        hols = Botswana(years=year)

def calc_hol2():
    hols = Botswana(years=range(1966, 2100))

t1 = timeit.timeit(calc_hol1, number=100)
t2 = timeit.timeit(calc_hol2, number=100)
```
| Test | Current beta |   PR  |
|:----:|:------------:|:-----:|
|  t1  |        2.740 | 2.379 |
|  t2  |       39.242 | 2.157 |